+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- The ability to run a user's own custom shell scripts placed in `./scripts/custom-command.sh`
 - Discord Integration. The user now has the ability to configure settings to post messages into their Discord server.
   - Unhandled exceptions. When the system runs into an unhandled error it will now post a message to your configured Discord server.
 - Password input is now hidden during setup